Polish #615, rename spring-cloud-alibaba-sentinel-zuul to spring-cloud-alibaba-sentinel-gateway
