+++ spring-cloud-alibaba-dependencies/pom.xml
@@ -268,7 +268,7 @@
             </dependency>
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
-                <artifactId>spring-cloud-alibaba-sentinel-zuul</artifactId>
+                <artifactId>spring-cloud-alibaba-sentinel-gateway</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
