+++ spring-cloud-alibaba-examples/sentinel-example/sentinel-zuul-example/pom.xml
@@ -40,7 +40,7 @@
 
         <dependency>
             <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-alibaba-sentinel-zuul</artifactId>
+            <artifactId>spring-cloud-alibaba-sentinel-gateway</artifactId>
         </dependency>
 
         <!--<dependency>-->
