+++ spring-cloud-alibaba-sentinel-gateway/README.md
@@ -14,7 +14,7 @@ Sentinel can provide `ServiceId` level and `API Path` level flow control for spr
 ```xml
 <dependency>
     <groupId>org.springframework.cloud</groupId>
-    <artifactId>spring-cloud-alibaba-sentinel-zuul</artifactId>
+    <artifactId>spring-cloud-alibaba-sentinel-gateway</artifactId>
     <version>x.y.z</version>
 </dependency>
