+++ spring-cloud-alibaba-sentinel-gateway/pom.xml
@@ -10,8 +10,8 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.springframework.cloud</groupId>
-    <artifactId>spring-cloud-alibaba-sentinel-zuul</artifactId>
-    <name>Spring Cloud Alibaba Sentinel Zuul</name>
+    <artifactId>spring-cloud-alibaba-sentinel-gateway</artifactId>
+    <name>Spring Cloud Alibaba Sentinel Gateway</name>
 
     <dependencies>
         <dependency>
